Use posixpath to construct url path
Fix for #8 (and #6?). Instead of manually joining the path, which may or
may not have extra slashes, use the posixpath module.

Found by way of @GP89's answer to
http://stackoverflow.com/questions/1793261/how-to-join-components-of-a-path-when-you-are-constructing-a-url-in-python

On same page, urlparse module is voted much higher, but there is a lot
of back and forth on handling leading slashes. Posix path just feels
cleaner to me, and it works just fine on Windows (wish I'd known of it
ages ago).